mybot/requirements.txt
Django>=4.2
python-telegram-bot==20.7
yt-dlp
python-dotenv
fastapi==0.111.2
uvicorn==0.24.1